refactor: PG 결제 Strategy 공통 추상 클래스 추출 (#60)
