+++ src/main/java/com/reservation/domain/payment/service/strategy/YPayPaymentStrategy.java
@@ -1,40 +1,18 @@
 package com.reservation.domain.payment.service.strategy;
 
-import com.reservation.domain.order.entity.Order;
-import com.reservation.domain.payment.entity.Payment;
 import com.reservation.domain.payment.entity.PaymentMethod;
 import com.reservation.domain.payment.pg.PgClient;
-import com.reservation.domain.payment.pg.PgPaymentResult;
-import com.reservation.global.exception.GeneralException;
-import com.reservation.global.exception.code.ErrorCode;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 @Component
-@RequiredArgsConstructor
-public class YPayPaymentStrategy implements PaymentStrategy {
+public class YPayPaymentStrategy extends PgPaymentStrategy {
 
-    private final PgClient pgClient;
+    public YPayPaymentStrategy(PgClient pgClient) {
+        super(pgClient);
+    }
 
     @Override
     public PaymentMethod getPaymentMethod() {
         return PaymentMethod.Y_PAY;
-    }
-
-    @Override
-    public void pay(Payment payment, Order order) {
-        PgPaymentResult result = pgClient.pay(PaymentMethod.Y_PAY, payment.getAmount());
-
-        if (!result.success()) {
-            payment.fail();
-            throw new GeneralException(ErrorCode.PAYMENT_FAILED);
-        }
-
-        payment.approve(result.transactionId());
-    }
-
-    @Override
-    public void cancel(Payment payment, Order order) {
-        pgClient.cancel(payment.getTransactionId());
     }
 }
